ordering is screwed for arrays since it attempts to do it all at the end

+++ def.go
@@ -94,7 +94,7 @@ type Field struct {
 	ValueType           string
 	Default             interface{}
 	DefaultType         string
-	Order               int
+	Order               float64
 	ReadOnly            bool
 	notTopLevel         bool
 	IsArrayChild        bool
